requirements.txt
streamlit
openai
python-dotenv
PyPDF2
scikit-learn
numpy
tiktoken
